+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -34,6 +34,8 @@ CHANGELOG
  * Register `RateLimiterBundle` automatically when the RateLimiter component is installed
  * Make `framework.webhook` an alias of the `webhook` configuration provided by `WebhookBundle`
  * Register `WebhookBundle` automatically when the Webhook component is installed
+ * Make `framework.http_client` an alias of the `http_client` configuration provided by `HttpClientBundle`
+ * Register `HttpClientBundle` automatically when the HttpClient component is installed
  * Make `framework.workflows` an alias of the `workflow` configuration provided by `WorkflowBundle`
  * Register `WorkflowBundle` automatically when the Workflow component is installed
  * Make `framework.remote_event` an alias of the `remote_event` configuration provided by `RemoteEventBundle`